Make a base OsuButton class to handle default samples
