+++ osu.Game.Tests/Visual/TestCaseReplaySettingsOverlay.cs
@@ -22,7 +22,7 @@ namespace osu.Game.Tests.Visual
 
             Add(container = new ExampleContainer());
 
-            AddStep(@"Add button", () => container.Add(new OsuButton
+            AddStep(@"Add button", () => container.Add(new TriangleButton
             {
                 RelativeSizeAxes = Axes.X,
                 Text = @"Button",
