+++ osu.Game/osu.Game.csproj
@@ -271,6 +271,7 @@
     <Compile Include="Beatmaps\Drawables\BeatmapSetHeader.cs" />
     <Compile Include="Database\DatabaseContextFactory.cs" />
     <Compile Include="Database\IHasPrimaryKey.cs" />
+    <Compile Include="Graphics\UserInterface\OsuButton.cs" />
     <Compile Include="Migrations\20171019041408_InitialCreate.cs" />
     <Compile Include="Migrations\20171019041408_InitialCreate.Designer.cs">
       <DependentUpon>20171019041408_InitialCreate.cs</DependentUpon>
@@ -366,7 +367,7 @@
     <Compile Include="Graphics\UserInterface\LoadingAnimation.cs" />
     <Compile Include="Graphics\UserInterface\MenuItemType.cs" />
     <Compile Include="Graphics\UserInterface\Nub.cs" />
-    <Compile Include="Graphics\UserInterface\OsuButton.cs" />
+    <Compile Include="Graphics\UserInterface\TriangleButton.cs" />
     <Compile Include="Graphics\UserInterface\OsuCheckbox.cs" />
     <Compile Include="Graphics\UserInterface\OsuContextMenu.cs" />
     <Compile Include="Graphics\UserInterface\OsuDropdown.cs" />
